Refactor: Update include paths and remove unused detail headers for cleaner structure
